refactor: resolve build warnings, reorg
