--- src/hooks/useMapFilters/schema.ts
@@ -1,11 +1,0 @@
-import { z } from 'zod';
-
-export const FiltersSchema = z.object({
-  'program-types': z.string().array(),
-  venues: z.string().array(),
-  audiences: z.string().array(),
-  'organization-name': z.string(),
-  address: z.string(),
-  regions: z.string().array(),
-  languages: z.string().array()
-});
